feat(project): Update LICENSE to include MIT License terms
The LICENSE file has been updated to include the full text of the MIT License. This change grants permission for use, modification, and distribution of the software, while clarifying that it is provided "as is" without any warranty. The copyright notice has also been updated to reflect the year 2025 and the entity "7irelo."